# backend/api/gen/go/app/service/v1/i_user_coupon_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: app/service/v1/i_user_coupon.proto

package servicev1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-shop/api/gen/go/coupon/service/v1"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationUserCouponServiceGet = "/app.service.v1.UserCouponService/Get"
const OperationUserCouponServiceList = "/app.service.v1.UserCouponService/List"
const OperationUserCouponServiceQuote = "/app.service.v1.UserCouponService/Quote"

type UserCouponServiceHTTPServer interface {
	Get(context.Context, *v11.GetUserCouponRequest) (*v11.UserCoupon, error)
	List(context.Context, *v1.PagingRequest) (*v11.ListUserCouponResponse, error)
	Quote(context.Context, *v11.QuoteRequest) (*v11.QuoteResponse, error)
}

func RegisterUserCouponServiceHTTPServer(s *http.Server, srv UserCouponServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/app/v1/mall/user-coupons", _UserCouponService_List17_HTTP_Handler(srv))
	r.GET("/app/v1/mall/user-coupons/{id}", _UserCouponService_Get17_HTTP_Handler(srv))
	r.POST("/app/v1/mall/user-coupons/quote", _UserCouponService_Quote0_HTTP_Handler(srv))
}

func _UserCouponService_List17_HTTP_Handler(srv UserCouponServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserCouponServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListUserCouponResponse)
		return ctx.Result(200, reply)
	}
}

func _UserCouponService_Get17_HTTP_Handler(srv UserCouponServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetUserCouponRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserCouponServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetUserCouponRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.UserCoupon)
		return ctx.Result(200, reply)
	}
}

func _UserCouponService_Quote0_HTTP_Handler(srv UserCouponServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.QuoteRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserCouponServiceQuote)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Quote(ctx, req.(*v11.QuoteRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.QuoteResponse)
		return ctx.Result(200, reply)
	}
}

type UserCouponServiceHTTPClient interface {
	Get(ctx context.Context, req *v11.GetUserCouponRequest, opts ...http.CallOption) (rsp *v11.UserCoupon, err error)
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListUserCouponResponse, err error)
	Quote(ctx context.Context, req *v11.QuoteRequest, opts ...http.CallOption) (rsp *v11.QuoteResponse, err error)
}

type UserCouponServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewUserCouponServiceHTTPClient(client *http.Client) UserCouponServiceHTTPClient {
	return &UserCouponServiceHTTPClientImpl{client}
}

func (c *UserCouponServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetUserCouponRequest, opts ...http.CallOption) (*v11.UserCoupon, error) {
	var out v11.UserCoupon
	pattern := "/app/v1/mall/user-coupons/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserCouponServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *UserCouponServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListUserCouponResponse, error) {
	var out v11.ListUserCouponResponse
	pattern := "/app/v1/mall/user-coupons"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserCouponServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *UserCouponServiceHTTPClientImpl) Quote(ctx context.Context, in *v11.QuoteRequest, opts ...http.CallOption) (*v11.QuoteResponse, error) {
	var out v11.QuoteResponse
	pattern := "/app/v1/mall/user-coupons/quote"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserCouponServiceQuote))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}