# backend/app/app/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Content Hub App API
    description: GoWind Content Hub App API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-shop
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-shop/blob/master/LICENSE
    version: "1.0"
paths:
...
    /app/v1/mall/user-coupons:
        get:
            tags:
                - UserCouponService
            operationId: UserCouponService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/coupon.service.v1.ListUserCouponResponse'
    /app/v1/mall/user-coupons/quote:
        post:
            tags:
                - UserCouponService
            operationId: UserCouponService_Quote
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/coupon.service.v1.QuoteRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/coupon.service.v1.QuoteResponse'
    /app/v1/mall/user-coupons/{id}:
        get:
            tags:
                - UserCouponService
            operationId: UserCouponService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/coupon.service.v1.UserCoupon'
...
components:
    schemas:
...
        coupon.service.v1.ListUserCouponResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/coupon.service.v1.UserCoupon'
                total:
                    type: string
        coupon.service.v1.QuoteRequest:
            type: object
            properties:
                userCouponId:
                    type: integer
                    description: 要试算的券实例ID
                    format: uint32
        coupon.service.v1.QuoteResponse:
            type: object
            properties:
                applicable:
                    type: boolean
                    description: 该券是否适用于当前购物车
                preDiscountTotal:
                    type: string
                    description: 折前总额（最小货币单位，分）
                discount:
                    type: string
                    description: 抵扣额（最小货币单位，分；不适用时为 0）
                postDiscountTotal:
                    type: string
                    description: 折后应付额（最小货币单位，分；= max(0, pre_discount_total - discount)）
        coupon.service.v1.UserCoupon:
            type: object
            properties:
                id:
                    type: integer
                    description: 券实例ID
                    format: uint32
                userId:
                    type: integer
                    description: 归属用户ID（由隐私层强制）
                    format: uint32
                couponTemplateId:
                    type: integer
                    description: 关联的优惠券模板ID
                    format: uint32
                status:
                    enum:
                        - STATUS_UNSPECIFIED
                        - UNUSED
                        - USED
                        - EXPIRED
                    type: string
                    description: 券实例状态
                    format: enum
                redeemedAt:
                    type: string
                    description: 核销时间（status=USED 时填写，返还时清空）
                    format: date-time
                redeemedOrderId:
                    type: integer
                    description: 核销时关联的订单ID（UNUSED 时为空）
                    format: uint32
                appliedDiscountAmount:
                    type: string
                    description: 实际抵扣额（最小货币单位，分；券侧审计，返还时清空）
                tenantId:
                    type: integer
                    description: 租户ID
                    format: uint32
                createdBy:
                    type: integer
                    description: 创建者用户ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者用户ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 用户优惠券实例
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...